move dark mode artifacts to services folder; remove non-applicable app comp test instance;
